+++ common/CHANGELOG.md
@@ -6,13 +6,13 @@ This document contains all changes to the crate since version 0.9.4.
 
 ### Breaking changes
 
-- The `Error::NotAscii` variant now stores the unencodable character instead of just the first byte of it.
+- `Error` is now a struct.
 - Removed the `byte` function from `Error`.
 
 ### Minor changes
 
-- Add `is_not_ascii` and `is_unencodable_ascii` variant checking functions to `Error`.
 - Add `char` function to `Error`.
+- Add `index` functio to `Error`.
 
 ## 0.10.4